model-builder generate-dockerfile all not to delete files in dockerfiles/ml

+++ tools/docker/assembler.py
@@ -1121,7 +1121,7 @@ def main(argv):
       try:
         if os.path.isfile(file_path):
           os.unlink(file_path)
-        elif os.path.isdir(file_path):
+        elif os.path.isdir(file_path) and file_path != 'models/dockerfiles/ml':
           shutil.rmtree(file_path)
       except Exception as e:
         print(e)
